Pets/interrior_segmentation/requirements.txt
tensorflow~=2.5.0
tensorflow-gpu~=2.5.0
numpy~=1.19.2
pillow~=9.0.1
scikit-learn~=1.0.2
uvicorn~=0.17.6
fastapi~=0.75.0
pydantic~=1.9.0
requests~=2.27.1